try to clean type of integral for for the fure and coorect pb of sabe mesh in float not in double of meshS and MeshL

+++ src/femlib/MeshLn.cpp
@@ -692,7 +692,7 @@ namespace Fem2D
         cerr <<"  -- MeshL**::Save  UNABLE TO OPEN  :"<< filename << endl;
         return(1);
     }
-    float fx,fy,fz;
+    double fx,fy,fz;
     // write vertice (meshL)
     GmfSetKwd(outm,GmfVertices,nv);
     for (int k=0; k<nv; k++) {
